voice_detection/requirements.txt
fastapi
uvicorn
python-multipart
numpy
librosa
soundfile
pydub
scipy
torch
torchaudio
transformers
xgboost
scikit-learn
joblib
redis
python-dotenv
httpx
